+++ src/apps/chifra/pkg/ledger/trail_balance.go
@@ -16,7 +16,7 @@ import (
 // function tries to price the asset. it then prints optional debugging information. Note that
 // the statement may be modified in this function.
 func (l *Ledger) trialBalance(reason types.TrialBalType, s *types.Statement) bool {
-	key := l.ctxKey(s.BlockNumber, s.TransactionIndex)
+	key := l.getAppContextKey(s.BlockNumber, s.TransactionIndex)
 	if ctx, exists := l.appContexts[key]; !exists {
 		logger.Fatal(fmt.Sprintf("should never happen - no context for %s", key))
